+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/SentryTasksProvider.kt
@@ -3,6 +3,7 @@ package io.sentry.android.gradle
 import com.android.build.gradle.api.ApplicationVariant
 import com.android.build.gradle.tasks.MergeSourceSetFolders
 import com.android.build.gradle.tasks.PackageAndroidArtifact
+import io.sentry.android.gradle.util.GroovyCompat.isDexguardAvailable
 import io.sentry.android.gradle.util.SentryPluginUtils.capitalizeUS
 import java.io.File
 import org.gradle.api.Project
@@ -87,16 +88,16 @@ internal object SentryTasksProvider {
     fun getMappingFileProvider(
         project: Project,
         variant: ApplicationVariant
-    ): Provider<FileCollection> = if (project.plugins.hasPlugin("com.guardsquare.proguard")) {
+    ): Provider<FileCollection> = if (project.plugins.hasPlugin("com.guardsquare.proguard") ||
+        isDexguardAvailable(project)
+    ) {
         val sep = File.separator
-        project.provider {
-            project.files(
-                File(
-                    project.buildDir,
-                    "outputs${sep}proguard${sep}${variant.name}${sep}mapping${sep}mapping.txt"
-                )
-            )
+        val path = if (project.plugins.hasPlugin("com.guardsquare.proguard")) {
+            "outputs${sep}proguard${sep}${variant.name}${sep}mapping${sep}mapping.txt"
+        } else {
+            "outputs${sep}dexguard${sep}mapping${sep}apk${sep}${variant.name}${sep}mapping.txt"
         }
+        project.provider { project.files(File(project.buildDir, path)) }
     } else {
         variant.mappingFileProvider
     }
@@ -122,5 +123,5 @@ internal object SentryTasksProvider {
             }
             .firstOrNull()
 
-    private val String.capitalized: String get() = this.capitalizeUS()
+    internal val String.capitalized: String get() = this.capitalizeUS()
 }